Move ScheduleStatus.kt to different package
